cpukit: RISC-V - make riscv32 code work for riscv64 - v2
* Use #ifdefs for 32/64 bit code
* Use unsigned long which is 32-bit on riscv32 and 64-bit on riscv64 (register size)
* Move the code to a new shared riscv folder to be shared between riscv32 and riscv64
* Rename RTEMS_CPU extracted from command line to shared riscv target s/riscv*/riscv

Update #3109

+++ cpukit/score/cpu/riscv/cpu.c
@@ -1,5 +1,5 @@
 /*
- * riscv32 CPU Dependent Source
+ * RISC-V CPU Dependent Source
  *
  * Copyright (c) 2015 University of York.
  * Hesham ALmatary <hesham@alumni.york.ac.uk>
@@ -59,12 +59,12 @@ void _CPU_Initialize(void)
   /* Do nothing */
 }
 
-void _CPU_ISR_Set_level(uint32_t level)
+void _CPU_ISR_Set_level(unsigned long level)
 {
   /* Do nothing */
 }
 
-uint32_t  _CPU_ISR_Get_level( void )
+unsigned long  _CPU_ISR_Get_level( void )
 {
   /* Do nothing */
   return 0;
@@ -80,7 +80,7 @@ void _CPU_ISR_install_raw_handler(
 }
 
 void _CPU_ISR_install_vector(
-  uint32_t    vector,
+  unsigned long    vector,
   proc_ptr    new_handler,
   proc_ptr   *old_handler
 )
